selftest: Replace bitbake -p with bitbake -e
Parsing all the recipes is annoying when trying to re-execute oe-selftest
and also unnecessary as its really just a sanity check. When the tests were
originally being developed the guard was useful but less so now.

Replace it with bitbake -e which is fast and checks the basic configuration
is valid.

(From OE-Core rev: 85dc8ca7633d8f247be160f8fa573465dd830f83)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>